Website Front End
This is the code used for making the website to access data. The three possible commands on the website are to Delete data inside the databases, Sort the output of the data by certain parameters (highest point scored, most wins, etc.), and to simply view the data (by ID number).